refactor metrics into seperate file, seperate plotting
