+++ Tools/ecl_ekf/detectors.py
@@ -1,3 +1,4 @@
+#! /usr/bin/env python3
 """
 detectors
 """
